Move introspection comparison to a dedicated module, for log filtering
Example usage:
`router --log error,apollo_router=info,apollo_router::query_planner::dual_introspection=trace`
